Use cloudinit inside of libvirt
Start using cloudinit when deploying to libvirt.

= The problem

In lots of places we have hard coded IPs. We want to move back to
resolve machiens by their FQDN. This is going to work on libvirt when
a local network is being used but it will work also when a bridge is
being used as long as the DNS of the external network can resolve the
leases of the DHCP server.

= Why cloudinit

This is what happens when a machine is created:

  * terraform creates the VM
  * the VM is booted
  * by default the hostname is "linux"
  * the OS will register against the local DHCP as "linux.k8s.local"
  * the VM will be pingable as "linux.k8s.local"
  * terraform provisioning steps are started as soon as it can ssh into
    the VM
  * one of our provisioning steps will change the hostname via
    "hostnamectl set-hostname foo.k8s.local"

Things seems fine now, however it's **not** possible to ping
`foo.k8s.local` because the machine is still known as "linux" by the
DNS. To fix the problem we would need to restart the network stack of
the VM. This will lead to troubles (timeouts, provisioning errors of
the subsequent scripts,...).

The right solution is to set the hostname **before** the network stack
is initialized. This can be done using cloud-init.

This commit introduces one cloud-init resource for each type of VM we
are going to create. Each VM will be provisioned with its final
hostname and will be immediately discoverable by the other ones.

= Issues

Cloud-init has been designed to be used in a cloud context. Here for
security reasons the root account is locked and can be accessed only
via ssh using valid keys.

By default cloud-init locks out of the machines. There are options to
prevent that but we are not exposing these directives inside of the
terraform-libvirt provider.

What does it mean: you cannot ssh into any box using root/vagrant.

What you can still do:

  * ssh/log in any machine as vagrant/vagrant user.
  * `sudo su -` as vagrant user
  * ssh as root into any machine using the ssh key you defined when
    provisioning the cluster

The terraform files have been changed to ssh into machines using the
right key (by default the passwordless `ssh/id_docker` inside of this
project).