Allow to create a minimal terraform setup
This will allow us to pass SKIP_DASHBOARD (completely skipping the
creation of a dashboard machine) and a SALT_MASTER_HOST (that will
allow us to choose a salt master where workers will connect to).

This will make the provisioning much less network dependant, and
much faster, since we already have the salt master, certificate authority,
and dashboard in the development environment that we already spawn
for development purposes.

In this context we will also ignore the creation of a "master"
kubernetes machine, since them all will be minions, and the master will
be randomly selected for now from the dashboard.