Make dashboard reachable when using the development environment.
When using the development environment we don't have a dedicated
dashboard machine. etcd discovery now relies on this dashboard hostname
by default, and we want to use the same salt states and pillar as in
production (there is a pillar setting the dashboard).

With this change, we will add a /etc/hosts entry pointing to the
"dashboard" machine, which is the development environment, only if
we specified the `DASHBOARD_HOST` envvar. This should not need any
kind of interaction from the user/developer perspective, it will just
do the right thing.